* commit '21c0a64351797011448e1bf45a81567ad6c5a5e8':
  refs #MSF-1668 Payment Order SQL and NoSQL codebase in line
  refs #MSF-1668 Payment Order SQL and NoSQL codebase in line
  refs #MSF-1668 Payment Order SQL and NoSQL codebase in line
  refs #MSF-1668 Payment Order SQL and NoSQL codebase in line

+++ msfwk-test/ms-payments-test/src/test/java/com/temenos/microservice/payments/api/test/CreateNewPaymentOrderITTest.java
@@ -18,6 +18,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.web.reactive.function.BodyInserters;
 import org.springframework.web.reactive.function.client.ClientResponse;
 
+import com.temenos.microservice.framework.core.conf.Environment;
 import com.temenos.microservice.framework.test.dao.Attribute;
 
 import reactor.core.publisher.Mono;
@@ -38,10 +39,14 @@ public class CreateNewPaymentOrderITTest extends ITTest {
 
 	@AfterClass
 	public static void clearData() {
-		deletePaymentOrderRecord("ms_payment_order", "paymentOrderId", "eq", "string", "PO~123~124~USD~100",
-				"debitAccount", "eq", "string", "123");
-		deletePaymentOrderRecord("ms_reference_data", "type", "eq", "string", "paymentref", "value", "eq", "string",
-				"PayRef");
+		if ("MYSQL".equals(Environment.getEnvironmentVariable("DB_VENDOR", ""))) {
+			clearRecords("PO~123~124~USD~100", "123");
+		} else {
+			deletePaymentOrderRecord("ms_payment_order", "paymentOrderId", "eq", "string", "PO~123~124~USD~100",
+					"debitAccount", "eq", "string", "123");
+			deletePaymentOrderRecord("ms_reference_data", "type", "eq", "string", "paymentref", "value", "eq", "string",
+					"PayRef");
+		}
 		daoFacade.closeConnection();
 	}
